final code.

+++ car.h
@@ -10,10 +10,9 @@ The library leverages the motor.h library to control the car.*/
 
 #include <stdint.h>
 #include "motor.h"
-#include "ultrasonic.h"
 
-void moveStraight(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t straightMode, unsigned int delayTime, uint8_t rightSpeed, uint8_t leftSpeed);
-void rotateInPlace(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t steeringMode, unsigned int delayTime, uint8_t speed);
+void moveStraight(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t straightMode, unsigned int delayTime, uint16_t rightSpeed, uint16_t leftSpeed);
+void rotateInPlace(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t steeringMode, unsigned int delayTime, uint16_t speed);
 void stopMoving(Motor_t *rightMotor, Motor_t *leftMotor);
 
 #endif*/
