+++ car.ino
@@ -7,7 +7,7 @@
   delayTime -> Decides the delay time (milliSeconds) the car is going to rotate in.
   rightSpeed -> The speed of the wheel on the right.
   leftSpeed -> The speed of the wheel on the left.*/
-void moveStraight(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t straightMode, unsigned int delayTime, uint8_t rightSpeed, uint8_t leftSpeed)
+void moveStraight(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t straightMode, unsigned int delayTime, uint16_t rightSpeed, uint16_t leftSpeed)
 {
   moveMotor(rightMotor, straightMode, rightSpeed);
   moveMotor(leftMotor, straightMode, leftSpeed);
@@ -19,7 +19,7 @@ void moveStraight(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t straightMode,
   leftMotor -> The wheel on the left of the car.
   steeringMode -> Decides whether the car is going to move RIGHT or LEFT.
   delayTime -> Decides the delay time the car is going to rotate in.*/
-void rotateInPlace(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t steeringMode, unsigned int delayTime, uint8_t speed)
+void rotateInPlace(Motor_t *rightMotor, Motor_t *leftMotor, uint8_t steeringMode, unsigned int delayTime, uint16_t speed)
 {
   moveMotor(rightMotor, (steeringMode + 1) % 2, speed);
   moveMotor(leftMotor, steeringMode % 2, speed);
